Refactor import paths for consistency and update cSpell configurations; add launch configuration for debugging main.py
