+++ .vscode/settings.json
@@ -1,16 +1,4 @@
 {
-  "cSpell.words": [
-    "casadi",
-    "njit",
-    "NMPC"
-  ],
-  "cSpell.ignoreWords": [
-    "F_pos"
-  ],
-  "python.analysis.extraPaths": ["./src"],
-  "python.autoComplete.extraPaths": ["./src"],
-  "python.envFile": "${workspaceFolder}/.env",
-  "python.testing.pytestArgs": ["test"],
-  "python.testing.unittestEnabled": false,
-  "python.testing.pytestEnabled": true
-}
+    "python.analysis.extraPaths": ["./src"],
+    "python.defaultInterpreterPath": "/opt/miniconda3/bin/python"
+  }
